Started with Raycasting. The ray direction and also the intersection with voxels are done. Trilinear interpolation and gradient computation for normals should be implemented.

+++ src/Volume.cpp
@@ -49,7 +49,7 @@ void Volume::compute_ddx_dddx()
 void Volume::zeroOutMemory()
 {
 	for (uint i1 = 0; i1 < dx * dy * dz; i1++) {
-		vol[i1] = double(0);
+		vol[i1] = double(1);
 		weights[i1] = double(0);
 		}
